+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 Major refactorization after two years, using TypeScript v3.
 
 ### Added
+- Support for replacement with instances of `Number`.
 - Share .vscode setup for launch, settings, and tasks in CVS.
 - .npmignore files, for distribution with minimal stuff.
 - Sync test for async operation.